refactor(execution-factory): review 后去重详情页基础信息接线
- 抽 DetailBasicInfoButton(按钮 + open 态 + 抽屉），四个详情页不再各自复制这套三件套，改渲染同一组件
- 抽 buildToolboxBasicInfoItems（utils），工具箱工具页与函数工作台共用；去掉只喂抽屉的死 span 元数据；函数工作台改用 useMemo（与兄弟一致）
- capability-ux 导出 HTTP_API_CAPABILITY_MODES，CreateMenu 与 ToolboxToolsScene 复用（消掉内联数组字面量，避免破坏向导 memo/effect 的稳定性）
- 函数工作台 subline 从内联硬编码色改为主题 .subline 类（var token，暗色安全）
- DetailBasicInfoDrawer 等宽字体走 var(--font-mono)
